Sending sms working

+++ app/src/main/java/com/example/myapplication/Models/Model.java
@@ -2,5 +2,5 @@ package com.example.myapplication.Models;
 
 public interface Model {
 
-    public void updateDt(int dt_ms);
+    public void updateDt(float dt_ms);
 }
